Move to ES Modules and bump hubot peer dependency
Hubot 4.0.0-alpha is actually only in existence in the Thesis branch
of hubot for now. Eventually this should probably become an
@thesis-co/hubot reference or a renamed package, but for now we stick to
what we have.

ES Modules require changes in hubot itself, and for hubot to support
ES Modules directly, its support for adapters is simplified to only
support ES Modules, thus the shift in hubot-matrix to ESM.